requirements.txt
Flask==3.0.2
gunicorn==21.2.0
pyswisseph==2.10.02